requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
openai
faiss-cpu
numpy
pypdf
tqdm
requests